fix(dracut-install): do not limit supplier handling to platform bus
The code that explores a sysfs node for suppliers looks for directories
with a "supplier:platform" prefix. Only the "supplier:" prefix refers to
the devlink relation itself here, the rest is the bus and device name of
the supplier. The bus is not necessarily "platform", but can be "phy",
"regulator", etc. Furthermore, we only have an initial search of devices
under "/sys/devices/platform" which also limits the dependencies we
check to the that directory.

Correct the searched prefix for sysfs supplier relations to "supplier:"
and the searched paths to "/sys/devices", so we can handle all supplier
dependencies to devices, not just those under "/sys/devices/platform".

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>