perf(dracut-install): use driver/module sysfs dirs for module name
To handle fw_devlink suppliers as dependencies, we traverse every device
in /sys and collect a hashmap of module names to their suppliers before
starting module installation work. The module names are found by reading
and passing the "modalias" file to the `kmod_module_new_from_lookup()`
function from libkmod. This function is unexpectedly slow on ARM64
systems with device-tree where this work is primarily necessary.

Usually (but not always), the running kernel has already looked up the
modalias and bound a driver from the associated module. Get module names
by resolving the "driver/module" symbolic link and taking its basename.
If it doesn't exist, fall back to looking up the modalias.

It's possible that multiple modules match the modalias, in which case
others will be ignored in favor of the one already chosen by the kernel.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>